+++ projects/rocshmem/src/gda/topology.cpp
@@ -50,7 +50,7 @@ namespace rocshmem
     std::vector<int> status(numPages);
 
     pages[0] = array;
-    for (int i = 1; i < numPages; i++) {
+    for (size_t i = 1; i < numPages; i++) {
       pages[i] = (char*)pages[i-1] + pageSize;
     }
 
@@ -897,7 +897,7 @@ namespace rocshmem
 #endif
 
           int minDistance = std::numeric_limits<int>::max();
-          for (int j = 0; j < ibvAddressList.size(); j++) {
+          for (size_t j = 0; j < ibvAddressList.size(); j++) {
             if (ibvAddressList[j] != "") {
               int distance = GetBusIdDistance(hipPciBusId, ibvAddressList[j]);
               if (distance < minDistance && distance >= 0) {
@@ -945,18 +945,18 @@ namespace rocshmem
 
     int numGpus = rocshmem::GetNumDevices(rocshmem::EXE_GPU);
     auto const& ibvDeviceList = rocshmem::GetIbvDeviceList();
-    for (int i = 0; i < ibvDeviceList.size(); i++) {
+    for (size_t i = 0; i < ibvDeviceList.size(); i++) {
 
       std::string closestGpusStr = "";
       for (int j = 0; j < numGpus; j++) {
-        if (rocshmem::GetClosestNicToGpu(j, nullptr, nullptr) == i) {
+        if (rocshmem::GetClosestNicToGpu(j, nullptr, nullptr) == static_cast<int>(i)) {
           if (closestGpusStr != "") closestGpusStr += ",";
           closestGpusStr += std::to_string(j);
         }
       }
 
       printf(" %-3d | %-11s | %-6s | %-12s | %-4d | %-14s | %-9s | %-20s\n",
-             i, ibvDeviceList[i].name.c_str(),
+             static_cast<int>(i), ibvDeviceList[i].name.c_str(),
              ibvDeviceList[i].hasActivePort ? "Yes" : "No",
              ibvDeviceList[i].busId.c_str(),
              ibvDeviceList[i].numaNode,
