+++ projects/rocshmem/src/rocshmem_gpu.cpp
@@ -596,7 +596,7 @@ __device__ void rocshmem_ctx_pe_quiet(rocshmem_ctx_t ctx, const int *target_pes,
 
   ContextTy *internal_ctx = get_internal_ctx(ctx);
 
-  for (int i = 0; i < npes;  i++) {
+  for (size_t i = 0; i < npes;  i++) {
     internal_ctx->pe_quiet(translate_pe(ctx, target_pes[i]));
   }
 }
