+++ projects/rocshmem/tests/unit_tests/atomic_wf_queue_gtest.hpp
@@ -108,8 +108,8 @@ class AtomicWFQueueTestFixture : public ::testing::Test {
     unsigned int expected_val {};
     unsigned int lane_id {};
     unsigned int idx {};
-    for (unsigned int i{0}; i < num_blocks; i++) {
-      for (unsigned int j{0}; j < block_size; j++) {
+    for (unsigned int i{0}; i < static_cast<unsigned int>(num_blocks); i++) {
+      for (unsigned int j{0}; j < static_cast<unsigned int>(block_size); j++) {
         idx = i * block_size + j;
         lane_id = j % wf_size;
         if (!lane_id) {
